refactor(frontend)!: fold v2/ into v1/ — one DDD tree
Merges the DI containers, drops every V2 name prefix (V2Record ->
SchemaRecord, V2RecordRepository -> SchemaRecordRepository), and deletes v2/.
