+++ extralit-frontend/pages/schemas/[id]/useSchemaRecordsViewModel.test.ts
@@ -2,18 +2,18 @@ import { beforeEach, describe, expect, it, vi } from "vitest";
 import { createPinia, setActivePinia } from "pinia";
 import Container from "ts-injecty";
 import { useResolveMock } from "~/v1/di/__mocks__/useResolveMock";
-import { GetSchemaRecordsUseCase } from "~/v2/domain/usecases/get-schema-records-use-case";
-import { SearchRecordsUseCase } from "~/v2/domain/usecases/search-records-use-case";
-import { GetSchemaSettingsUseCase } from "~/v2/domain/usecases/get-schema-settings-use-case";
-import { RecordsPage } from "~/v2/domain/entities/record/RecordsPage";
-import { V2Record } from "~/v2/domain/entities/record/V2Record";
-import { Schema } from "~/v2/domain/entities/schema/Schema";
-import { SchemaVersion } from "~/v2/domain/entities/schema/SchemaVersion";
-import { ColumnMeta } from "~/v2/domain/entities/schema/ColumnMeta";
-import { SearchCriteria } from "~/v2/domain/entities/search/SearchCriteria";
+import { GetSchemaRecordsUseCase } from "~/v1/domain/usecases/get-schema-records-use-case";
+import { SearchRecordsUseCase } from "~/v1/domain/usecases/search-records-use-case";
+import { GetSchemaSettingsUseCase } from "~/v1/domain/usecases/get-schema-settings-use-case";
+import { RecordsPage } from "~/v1/domain/entities/schema/RecordsPage";
+import { SchemaRecord } from "~/v1/domain/entities/schema/SchemaRecord";
+import { Schema } from "~/v1/domain/entities/schema/Schema";
+import { SchemaVersion } from "~/v1/domain/entities/schema/SchemaVersion";
+import { ColumnMeta } from "~/v1/domain/entities/schema/ColumnMeta";
+import { SearchCriteria } from "~/v1/domain/entities/search/SearchCriteria";
 import { useSchemaRecordsViewModel } from "./useSchemaRecordsViewModel";
 
-const RECORD = new V2Record("r-1", "s-1", "10.1000/j.x", null, { title: "A study" }, null, "pending", "", "");
+const RECORD = new SchemaRecord("r-1", "s-1", "10.1000/j.x", null, { title: "A study" }, null, "pending", "", "");
 const SETTINGS = {
   schema: new Schema("s-1", "sample_size", "ready", "w-1", "v-1", {}, "", ""),
   versions: [new SchemaVersion("v-1", "s-1", 1, "")],
