+++ extralit-frontend/pages/schemas/[id]/useSchemaRecordsViewModel.ts
@@ -1,13 +1,13 @@
 import { computed, onBeforeMount, ref } from "vue";
 import { useResolve } from "ts-injecty";
-import { GetSchemaRecordsUseCase } from "~/v2/domain/usecases/get-schema-records-use-case";
-import { SearchRecordsUseCase } from "~/v2/domain/usecases/search-records-use-case";
-import { GetSchemaSettingsUseCase } from "~/v2/domain/usecases/get-schema-settings-use-case";
-import { RecordsPage } from "~/v2/domain/entities/record/RecordsPage";
-import { SearchCriteria, type RecordFilter } from "~/v2/domain/entities/search/SearchCriteria";
-import { Schema } from "~/v2/domain/entities/schema/Schema";
-import { ColumnMeta } from "~/v2/domain/entities/schema/ColumnMeta";
-import { type V2RecordStatus } from "~/v2/domain/entities/record/V2Record";
+import { GetSchemaRecordsUseCase } from "~/v1/domain/usecases/get-schema-records-use-case";
+import { SearchRecordsUseCase } from "~/v1/domain/usecases/search-records-use-case";
+import { GetSchemaSettingsUseCase } from "~/v1/domain/usecases/get-schema-settings-use-case";
+import { RecordsPage } from "~/v1/domain/entities/schema/RecordsPage";
+import { SearchCriteria, type RecordFilter } from "~/v1/domain/entities/search/SearchCriteria";
+import { Schema } from "~/v1/domain/entities/schema/Schema";
+import { ColumnMeta } from "~/v1/domain/entities/schema/ColumnMeta";
+import { type SchemaRecordStatus } from "~/v1/domain/entities/schema/SchemaRecord";
 
 const PAGE_SIZE = 25;
 
@@ -21,7 +21,7 @@ export const useSchemaRecordsViewModel = (schemaId: string) => {
   const page = ref<RecordsPage>(new RecordsPage([], 0));
   const isLoading = ref(false);
   const searchText = ref("");
-  const statusFilter = ref<V2RecordStatus | "">("");
+  const statusFilter = ref<SchemaRecordStatus | "">("");
   const currentOffset = ref(0);
 
   const hasQuery = computed(() => Boolean(searchText.value.trim() || statusFilter.value));
