+++ extralit-frontend/pages/schemas/index.test.ts
@@ -4,8 +4,8 @@ import { ref } from "vue";
 import { createPinia, setActivePinia } from "pinia";
 import Container from "ts-injecty";
 import { useResolveMock } from "~/v1/di/__mocks__/useResolveMock";
-import { GetSchemasUseCase } from "~/v2/domain/usecases/get-schemas-use-case";
-import { Schema } from "~/v2/domain/entities/schema/Schema";
+import { GetSchemasUseCase } from "~/v1/domain/usecases/get-schemas-use-case";
+import { Schema } from "~/v1/domain/entities/schema/Schema";
 import { useWorkspaces } from "~/v1/infrastructure/storage/WorkspaceStorage";
 import { Workspace } from "~/v1/domain/entities/workspace/Workspace";
 import * as useEnsureWorkspacesModule from "~/composables/useEnsureWorkspaces";
