+++ extralit-frontend/v1/domain/entities/projection/WorkspaceProjection.ts
