+++ extralit-frontend/v1/domain/entities/projection/grid-adapter.ts
