+++ extralit-frontend/v1/domain/entities/schema/ColumnMeta.ts
