+++ extralit-frontend/v1/domain/entities/schema/RecordsPage.ts
@@ -1,8 +1,8 @@
-import { type V2Record } from "./V2Record";
+import { type SchemaRecord } from "./SchemaRecord";
 
 export class RecordsPage {
   constructor(
-    public readonly items: V2Record[],
+    public readonly items: SchemaRecord[],
     // Authoritative: v1's Elasticsearch-backed list/search endpoints return an exact count.
     public readonly total: number
   ) {}
