+++ extralit-frontend/v1/domain/entities/schema/Schema.ts
