+++ extralit-frontend/v1/domain/entities/search/SearchCriteria.test.ts
