+++ extralit-frontend/v1/domain/usecases/get-schemas-use-case.test.ts
@@ -2,7 +2,7 @@ import { beforeEach, describe, expect, it, vi } from "vitest";
 import { createPinia, setActivePinia } from "pinia";
 import { GetSchemasUseCase } from "./get-schemas-use-case";
 import { Schema } from "../entities/schema/Schema";
-import { useSchemas } from "~/v2/infrastructure/storage/SchemasStorage";
+import { useSchemas } from "~/v1/infrastructure/storage/SchemasStorage";
 
 const SCHEMA = new Schema("s-1", "sample_size", "published", "w-1", "v-1", {}, "2026-01-01", "2026-01-01");
