+++ extralit-frontend/v1/domain/usecases/search-records-use-case.ts
@@ -1,9 +1,9 @@
-import { RecordsPage } from "../entities/record/RecordsPage";
+import { RecordsPage } from "../entities/schema/RecordsPage";
 import { SearchCriteria } from "../entities/search/SearchCriteria";
-import { V2RecordRepository } from "~/v2/infrastructure/repositories/V2RecordRepository";
+import { SchemaRecordRepository } from "~/v1/infrastructure/repositories/SchemaRecordRepository";
 
 export class SearchRecordsUseCase {
-  constructor(private readonly recordRepository: V2RecordRepository) {}
+  constructor(private readonly recordRepository: SchemaRecordRepository) {}
 
   execute(schemaId: string, criteria: SearchCriteria): Promise<RecordsPage> {
     return this.recordRepository.searchRecords(schemaId, criteria);
