+++ extralit-frontend/v1/infrastructure/repositories/ProjectionRepository.ts
