--- extralit-frontend/v2/di/index.ts
@@ -1,1 +1,0 @@
-export * from "./di";
